pyproject.toml
[project]
name = "test-task-renovate"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2026.3",
    "django==5.2",
    "django-allauth>=65.14.3",
    "django-crispy-forms>=2.6",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.0",
    "pre-commit>=4.5.1",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-django>=4.5.0",
    "ruff>=0.15.4",
]